--- python/benchmark/images/generic_2d_all/scores_summary_best.csv
@@ -1,10 +1,0 @@
-method, PSNR, SSIM, MI, SMI
-blurry, 23.1, 0.77, 0.17, 0.38
-blurry\&noisy (input), 17.8, 0.29, 0.07, 0.18
-Conjugate Gradient TV, 19.4, 0.41, 0.09, 0.21
-Chambole Pock TV , 18.7, 0.40, 0.07, 0.23
-Lucy Richardson $n=5$, 22.2, 0.59, 0.12, 0.25
-Lucy Richardson $n=10$, 21.1, 0.52, 0.10, 0.25
-Lucy Richardson $n=20$, 18.5, 0.38, 0.08, 0.19
-SSI UNet \emph{no masking}, 17.7, 0.38, 0.07, 0.14
-SSI UNet, \textbf{22.5}, \textbf{0.61}, \textbf{0.14}, \textbf{0.27}
